fix: updated test suite to use `workspaceManager` and `Analyzer.initialzie()`
