+++ test-data/startup-workspace.test.ts
@@ -1,7 +1,8 @@
 // import * as fs from 'fs';
 import * as os from 'os';
 import { setLogger } from './helpers';
-import { FishUriWorkspace, initializeDefaultFishWorkspaces, workspaces } from '../src/utils/workspace';
+import { FishUriWorkspace, initializeDefaultFishWorkspaces } from '../src/utils/workspace';
+import { workspaceManager } from '../src/utils/workspace-manager';
 import { Config, config, ConfigSchema } from '../src/config';
 import { uriToPath } from '../src/utils/translation';
 import * as LSP from 'vscode-languageserver';
@@ -35,9 +36,7 @@ describe('setup workspace', () => {
 
   afterEach(() => {
     config.fish_lsp_all_indexed_paths = [];
-    for (const _ of workspaces) {
-      workspaces.pop();
-    }
+    workspaceManager.clear();
   });
 
   describe('fisher workspace', () => {
@@ -125,7 +124,7 @@ describe('setup workspace', () => {
       config.fish_lsp_single_workspace_support = false;
       const uri = `file://${os.homedir()}/.config/fish`;
       const workspaces = await initializeDefaultFishWorkspaces(uri);
-      expect(workspaces.length).toBe(2);
+      expect(workspaces.length).toBe(3);
       expect(config.fish_lsp_single_workspace_support).toBe(false);
     });
